debugging

+++ database/migrations/2018_05_28_114447_create_Users_table.php
@@ -15,7 +15,7 @@ class CreateUsersTable extends Migration {
 			$table->string('username', 20)->unique();
 			$table->string('password', 255);
 			$table->string('Telepon', 13);
-			$table->float('gaji');
+			$table->double('gaji');
             $table->rememberToken();
 			$table->text('Alamat');
 			$table->integer('cabang_id')->unsigned()->nullable();
